requirements.txt
Django==1.9.8
django-ckeditor==5.0.3
mysqlclient==1.3.7
Pillow==3.3.0
PyMySQL==0.7.6
webencodings==0.5
